+++ src/Task/Composer.php
@@ -2,6 +2,8 @@
 namespace Robo\Task;
 
 use Robo\Result;
+use Robo\Task\Shared\TaskException;
+use Robo\Task\Shared\TaskInterface;
 
 /**
  * Contains tasks for composer.